requirements.txt
# Dependencias del generador de facturas
requests>=2.32.0
python-dotenv>=1.0.0
beautifulsoup4>=4.14.0
flask>=3.0.0
flask-cors>=4.0.0
